Move statistics subsystem out of vabackend.c into src/stats.c
vabackend.c is already very large; the NVD_STATS counters, the periodic/
final dump and the live backing-image accounting are self-contained and
don't belong in the core dispatch file.

Move the NVStatCounter enum and the nvStats* API into a new src/stats.{c,h}
and pull the NVD_STATS env parsing into nvStatsInit(). The two driver
internals the dump needs - the thread id helper and the resolved stats
output stream - are now exposed as nv_gettid() and nvStatsOutput(). No
behavioural change; the counters and output format are identical.

(cherry picked from commit 609c6ce9d5240badc0f9667dee3412c7ef430574)
